server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.10.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
pandas==2.2.3
numpy==2.1.2
python-multipart==0.0.12
websockets==13.1
apscheduler==3.10.4
beanie==1.27.0